requirements.txt
matplotlib
numpy
pynvml
pyyaml
scipy
seaborn
tensorboardx
torchmetrics
transformers
tqdm